Bringing sweeteners back into main repo
The idea of having them seperate was so that people could extend the
language using the Base classes from the sweeteners. It was kind of a
pain in the ass for developing new ones.

Until the language is at a stable point, it's easier for them to live
inside this project.