Adding function blocks
Passing many annonymous functions as parameters to a function gets
messy, fast. I prefer the Promise route of using normal message passing
with a single anon-func at a time.

Creamerscript will have an (informal) convention that you should only
pass a single annonymous function to a method, a la Ruby's blocks with
Enumerable methods, for example. The syntax is a mix of Ruby and
Smalltalk:

    (people map) { person | person.name }